Update bigdata.md
Made headers consistent. Spelling. Created numbered list for first section -- can change that back if you like, since it is within a quote. But the line breaks in the code made it look like that was what you wanted to have.

Added colons to separate links from bullet statements in checklists area.